Move xarray tests into a single file
